+++ NN/main_TF1.0.py
@@ -6,7 +6,7 @@ import tensorflow as tf
 
 
 
-#### 모델 셋팅 시작 ####
+#### 모델 셋팅 시작 Computational Graph ####
 
 x = tf.placeholder(tf.float32, [None, 784])  # 데이터 담을 placeholder 선언
 
@@ -21,7 +21,7 @@ cross_entropy = tf.reduce_mean(-tf.reduce_sum(y_ * tf.log(y), axis=[1])) # Loss
 
 train_step = tf.train.GradientDescentOptimizer(0.5).minimize(cross_entropy)
 
-#### 모델 셋팅 끝 ####
+#### 모델 셋팅 끝 Computational Graph ####
 
 
 init = tf.global_variables_initializer() # 변수 초기화(텐서플로우 필수과정)
@@ -34,11 +34,13 @@ for i in range(1000):
   sess.run(train_step, feed_dict={x: batch_xs, y_: batch_ys})
 
 # print(batch_xs[0]) # 들어간 데이터 확인
-#### 테스팅 모델 셋팅 시작 ####
+#### 테스트 모델 셋팅 시작 ####
 correct_prediction = tf.equal(tf.argmax(y, 1), tf.argmax(y_, 1)) # argmax(input,  dimension of the input Tensor to reduce across)
 accuracy = tf.reduce_mean(tf.cast(correct_prediction, tf.float32))
-#### 테스팅 모델 셋팅 끝 ####
+#### 테스트 모델 셋팅 끝 ####
 print(mnist.test.labels)
+
+#### 테스트 모델 Operation ####
 print(sess.run(accuracy, feed_dict={x: mnist.test.images, y_: mnist.test.labels}))
