+++ README.md
@@ -1,15 +1,23 @@
 # DeepLearningTutorials
 본 문서는 Deep Learning 실습을 위한 Tensorflow 예제코드를 모아놓은 것 입니다.
 
-###16.12.8
+### 16.12.8
 MNIST 데이터 분석을 위한 NN, CNN 모델 예제
 Visualization 툴인 Tensorboard 예제
 Model 저장을 위한 Saver예제 포함
 
-###16.12.9
+### 16.12.9
 py -> ipynb 코드 변환,
 saver 예제는 해당 Directory에 save 폴더가 미리 있어야함
 
 
-###16.12.16
+### 16.12.16
 CNN Saver, restore 예제 추가
+
+### 17.3.28
+TF 1.0 Version에 맞게 코드 변경
+
+## Future work
+TF Embedding Visualization
+RNN
+Attention Implementation
